+++ templates/badgr-server-status.tpl.php
@@ -1,0 +1,8 @@
+<div class="cmb-row table-layout">
+	<div class="cmb-th">
+		<label for="badgr_server_status">Server Status</label>
+	</div>
+	<div class="cmb-td">
+		<?php echo \BadgeFactor2\BadgrClient::get_status(); ?>
+	</div>
+</div>