+++ CHANGELOG.md
@@ -10,4 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.0-alpha]
 ### Added
 - LICENSE, README and CHANGELOG files
-- Plugin skeleton created
+- Internal dependencies: CMB2, thephpleague/oauth2-client, guzzlehttp/guzzle; 
+- Badgr Client & Server settings page
+- Custom Post Types: Issuer, Badge, Assertion
+- WordPress CLI command